% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column.R
\name{Column}
\alias{Column}
\title{Column}
\value{
Schema class instance
}
\description{
Column class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data_type}}{\link{character} One of integer, numeric, character, or logical.}

\item{\code{allow_na}}{\link{logical} Should NAs be allowed in this column?}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Column$new()}}
\item \href{#method-clone}{\code{Column$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializes schema class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Column$new(data_type, allow_na = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_type}}{\link{character} One of integer, numeric, character, or logical.}

\item{\code{allow_na}}{\link{logical} Should NAs be allowed in this column?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Column$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
