% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\name{correspondsToSchema}
\alias{correspondsToSchema}
\title{Does the DF has a good schema?}
\usage{
correspondsToSchema(df, schema)
}
\arguments{
\item{df}{data.frame you want to check}

\item{schema}{list with names and r data types. Type: schema_example
in the console to see example schema}
}
\value{
TRUE/FALSE (is it ok or is not)
}
\description{
Does the DF has a good schema?
}
\examples{

}
