% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Demand.R
\name{Demand}
\alias{Demand}
\title{Demand}
\usage{
Demand(id = NULL, warranty = NULL, validThrough = NULL,
  validFrom = NULL, sku = NULL, serialNumber = NULL, seller = NULL,
  priceSpecification = NULL, mpn = NULL, itemOffered = NULL,
  itemCondition = NULL, inventoryLevel = NULL, ineligibleRegion = NULL,
  includesObject = NULL, gtin8 = NULL, gtin14 = NULL, gtin13 = NULL,
  gtin12 = NULL, eligibleTransactionVolume = NULL, eligibleRegion = NULL,
  eligibleQuantity = NULL, eligibleDuration = NULL,
  eligibleCustomerType = NULL, deliveryLeadTime = NULL,
  businessFunction = NULL, availableDeliveryMethod = NULL,
  availableAtOrFrom = NULL, availabilityStarts = NULL,
  availabilityEnds = NULL, availability = NULL, areaServed = NULL,
  advanceBookingRequirement = NULL, acceptedPaymentMethod = NULL,
  url = NULL, sameAs = NULL, potentialAction = NULL, name = NULL,
  mainEntityOfPage = NULL, image = NULL, identifier = NULL,
  disambiguatingDescription = NULL, description = NULL,
  alternateName = NULL, additionalType = NULL)
}
\arguments{
\item{id}{identifier for the object (URI)}

\item{warranty}{(WarrantyPromise or WarrantyPromise type.) The warranty promise(s) included in the offer.}

\item{validThrough}{(DateTime or DateTime or DateTime or DateTime or DateTime or DateTime or DateTime type.) The date after when the item is not valid. For example the end of an offer, salary period, or a period of opening hours.}

\item{validFrom}{(DateTime or DateTime or DateTime or DateTime or DateTime or DateTime or DateTime type.) The date when the item becomes valid.}

\item{sku}{(Text or Text or Text type.) The Stock Keeping Unit (SKU), i.e. a merchant-specific identifier for a product or service, or the product to which the offer refers.}

\item{serialNumber}{(Text or Text or Text type.) The serial number or any alphanumeric identifier of a particular product. When attached to an offer, it is a shortcut for the serial number of the product included in the offer.}

\item{seller}{(Person or Organization or Person or Organization or Person or Organization or Person or Organization or Person or Organization type.) An entity which offers (sells / leases / lends / loans) the services / goods.  A seller may also be a provider.}

\item{priceSpecification}{(PriceSpecification or PriceSpecification or PriceSpecification type.) One or more detailed price specifications, indicating the unit price and delivery or payment charges.}

\item{mpn}{(Text or Text or Text type.) The Manufacturer Part Number (MPN) of the product, or the product to which the offer refers.}

\item{itemOffered}{(Service or Product or Service or Product type.) The item being offered.}

\item{itemCondition}{(OfferItemCondition or OfferItemCondition or OfferItemCondition type.) A predefined value from OfferItemCondition or a textual description of the condition of the product or service, or the products or services included in the offer.}

\item{inventoryLevel}{(QuantitativeValue or QuantitativeValue or QuantitativeValue type.) The current approximate inventory level for the item or items.}

\item{ineligibleRegion}{(Text or Place or GeoShape or Text or Place or GeoShape or Text or Place or GeoShape type.) The ISO 3166-1 (ISO 3166-1 alpha-2) or ISO 3166-2 code, the place, or the GeoShape for the geo-political region(s) for which the offer or delivery charge specification is not valid, e.g. a region where the transaction is not allowed.See also [[eligibleRegion]].}

\item{includesObject}{(TypeAndQuantityNode or TypeAndQuantityNode type.) This links to a node or nodes indicating the exact quantity of the products included in the offer.}

\item{gtin8}{(Text or Text or Text type.) The [GTIN-8](http://apps.gs1.org/GDD/glossary/Pages/GTIN-8.aspx) code of the product, or the product to which the offer refers. This code is also known as EAN/UCC-8 or 8-digit EAN. See [GS1 GTIN Summary](http://www.gs1.org/barcodes/technical/idkeys/gtin) for more details.}

\item{gtin14}{(Text or Text or Text type.) The [GTIN-14](http://apps.gs1.org/GDD/glossary/Pages/GTIN-14.aspx) code of the product, or the product to which the offer refers. See [GS1 GTIN Summary](http://www.gs1.org/barcodes/technical/idkeys/gtin) for more details.}

\item{gtin13}{(Text or Text or Text type.) The [GTIN-13](http://apps.gs1.org/GDD/glossary/Pages/GTIN-13.aspx) code of the product, or the product to which the offer refers. This is equivalent to 13-digit ISBN codes and EAN UCC-13. Former 12-digit UPC codes can be converted into a GTIN-13 code by simply adding a preceeding zero. See [GS1 GTIN Summary](http://www.gs1.org/barcodes/technical/idkeys/gtin) for more details.}

\item{gtin12}{(Text or Text or Text type.) The [GTIN-12](http://apps.gs1.org/GDD/glossary/Pages/GTIN-12.aspx) code of the product, or the product to which the offer refers. The GTIN-12 is the 12-digit GS1 Identification Key composed of a U.P.C. Company Prefix, Item Reference, and Check Digit used to identify trade items. See [GS1 GTIN Summary](http://www.gs1.org/barcodes/technical/idkeys/gtin) for more details.}

\item{eligibleTransactionVolume}{(PriceSpecification or PriceSpecification or PriceSpecification type.) The transaction volume, in a monetary unit, for which the offer or price specification is valid, e.g. for indicating a minimal purchasing volume, to express free shipping above a certain order volume, or to limit the acceptance of credit cards to purchases to a certain minimal amount.}

\item{eligibleRegion}{(Text or Place or GeoShape or Text or Place or GeoShape or Text or Place or GeoShape type.) The ISO 3166-1 (ISO 3166-1 alpha-2) or ISO 3166-2 code, the place, or the GeoShape for the geo-political region(s) for which the offer or delivery charge specification is valid.See also [[ineligibleRegion]].}

\item{eligibleQuantity}{(QuantitativeValue or QuantitativeValue or QuantitativeValue type.) The interval and unit of measurement of ordering quantities for which the offer or price specification is valid. This allows e.g. specifying that a certain freight charge is valid only for a certain quantity.}

\item{eligibleDuration}{(QuantitativeValue or QuantitativeValue type.) The duration for which the given offer is valid.}

\item{eligibleCustomerType}{(BusinessEntityType or BusinessEntityType type.) The type(s) of customers for which the given offer is valid.}

\item{deliveryLeadTime}{(QuantitativeValue or QuantitativeValue type.) The typical delay between the receipt of the order and the goods either leaving the warehouse or being prepared for pickup, in case the delivery method is on site pickup.}

\item{businessFunction}{(BusinessFunction or BusinessFunction or BusinessFunction type.) The business function (e.g. sell, lease, repair, dispose) of the offer or component of a bundle (TypeAndQuantityNode). The default is http://purl.org/goodrelations/v1#Sell.}

\item{availableDeliveryMethod}{(DeliveryMethod or DeliveryMethod type.) The delivery method(s) available for this offer.}

\item{availableAtOrFrom}{(Place or Place type.) The place(s) from which the offer can be obtained (e.g. store locations).}

\item{availabilityStarts}{(DateTime or DateTime type.) The beginning of the availability of the product or service included in the offer.}

\item{availabilityEnds}{(DateTime or DateTime type.) The end of the availability of the product or service included in the offer.}

\item{availability}{(ItemAvailability or ItemAvailability type.) The availability of this item&#x2014;for example In stock, Out of stock, Pre-order, etc.}

\item{areaServed}{(Text or Place or GeoShape or AdministrativeArea or Text or Place or GeoShape or AdministrativeArea or Text or Place or GeoShape or AdministrativeArea or Text or Place or GeoShape or AdministrativeArea or Text or Place or GeoShape or AdministrativeArea or Text or Place or GeoShape or AdministrativeArea type.) The geographic area where a service or offered item is provided.}

\item{advanceBookingRequirement}{(QuantitativeValue or QuantitativeValue type.) The amount of time that is required between accepting the offer and the actual usage of the resource or service.}

\item{acceptedPaymentMethod}{(PaymentMethod or LoanOrCredit or PaymentMethod or LoanOrCredit type.) The payment method(s) accepted by seller for this offer.}

\item{url}{(URL type.) URL of the item.}

\item{sameAs}{(URL type.) URL of a reference Web page that unambiguously indicates the item's identity. E.g. the URL of the item's Wikipedia page, Wikidata entry, or official website.}

\item{potentialAction}{(Action type.) Indicates a potential Action, which describes an idealized action in which this thing would play an 'object' role.}

\item{name}{(Text type.) The name of the item.}

\item{mainEntityOfPage}{(URL or CreativeWork type.) Indicates a page (or other CreativeWork) for which this thing is the main entity being described. See [background notes](/docs/datamodel.html#mainEntityBackground) for details.}

\item{image}{(URL or ImageObject type.) An image of the item. This can be a [[URL]] or a fully described [[ImageObject]].}

\item{identifier}{(URL or Text or PropertyValue type.) The identifier property represents any kind of identifier for any kind of [[Thing]], such as ISBNs, GTIN codes, UUIDs etc. Schema.org provides dedicated properties for representing many of these, either as textual strings or as URL (URI) links. See [background notes](/docs/datamodel.html#identifierBg) for more details.}

\item{disambiguatingDescription}{(Text type.) A sub property of description. A short description of the item used to disambiguate from other, similar items. Information from other properties (in particular, name) may be necessary for the description to be useful for disambiguation.}

\item{description}{(Text type.) A description of the item.}

\item{alternateName}{(Text type.) An alias for the item.}

\item{additionalType}{(URL type.) An additional type for the item, typically used for adding more specific types from external vocabularies in microdata syntax. This is a relationship between something and a class that the thing is in. In RDFa syntax, it is better to use the native RDFa syntax - the 'typeof' attribute - for multiple types. Schema.org tools may have only weaker understanding of extra types, in particular those defined externally.}
}
\value{
a list object corresponding to a schema:Demand
}
\description{
A demand entity represents the public, not necessarily binding, not necessarily exclusive, announcement by an organization or person to seek a certain type of goods or services. For describing demand using this type, the very same properties used for Offer apply.
}
