% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EducationalOrganization.R
\name{EducationalOrganization}
\alias{EducationalOrganization}
\title{EducationalOrganization}
\usage{
EducationalOrganization(id = NULL, alumni = NULL, vatID = NULL,
  telephone = NULL, taxID = NULL, subOrganization = NULL,
  sponsor = NULL, serviceArea = NULL, seeks = NULL, reviews = NULL,
  review = NULL, publishingPrinciples = NULL, parentOrganization = NULL,
  owns = NULL, numberOfEmployees = NULL, naics = NULL, members = NULL,
  memberOf = NULL, member = NULL, makesOffer = NULL, logo = NULL,
  location = NULL, leiCode = NULL, legalName = NULL, isicV4 = NULL,
  hasPOS = NULL, hasOfferCatalog = NULL, globalLocationNumber = NULL,
  funder = NULL, foundingLocation = NULL, foundingDate = NULL,
  founders = NULL, founder = NULL, faxNumber = NULL, events = NULL,
  event = NULL, employees = NULL, employee = NULL, email = NULL,
  duns = NULL, dissolutionDate = NULL, department = NULL,
  contactPoints = NULL, contactPoint = NULL, brand = NULL,
  awards = NULL, award = NULL, areaServed = NULL,
  aggregateRating = NULL, address = NULL, url = NULL, sameAs = NULL,
  potentialAction = NULL, name = NULL, mainEntityOfPage = NULL,
  image = NULL, identifier = NULL, disambiguatingDescription = NULL,
  description = NULL, alternateName = NULL, additionalType = NULL)
}
\arguments{
\item{id}{identifier for the object (URI)}

\item{alumni}{(Person or Person type.) Alumni of an organization.}

\item{vatID}{(Text or Text type.) The Value-added Tax ID of the organization or person.}

\item{telephone}{(Text or Text or Text or Text type.) The telephone number.}

\item{taxID}{(Text or Text type.) The Tax / Fiscal ID of the organization or person, e.g. the TIN in the US or the CIF/NIF in Spain.}

\item{subOrganization}{(Organization type.) A relationship between two organizations where the first includes the second, e.g., as a subsidiary. See also: the more specific 'department' property.}

\item{sponsor}{(Person or Organization or Person or Organization or Person or Organization or Person or Organization type.) A person or organization that supports a thing through a pledge, promise, or financial contribution. e.g. a sponsor of a Medical Study or a corporate sponsor of an event.}

\item{serviceArea}{(Place or GeoShape or AdministrativeArea or Place or GeoShape or AdministrativeArea or Place or GeoShape or AdministrativeArea type.) The geographic area where the service is provided.}

\item{seeks}{(Demand or Demand type.) A pointer to products or services sought by the organization or person (demand).}

\item{reviews}{(Review or Review or Review or Review or Review type.) Review of the item.}

\item{review}{(Review or Review or Review or Review or Review or Review or Review or Review type.) A review of the item.}

\item{publishingPrinciples}{(URL or CreativeWork or URL or CreativeWork or URL or CreativeWork type.) The publishingPrinciples property indicates (typically via [[URL]]) a document describing the editorial principles of an [[Organization]] (or individual e.g. a [[Person]] writing a blog) that relate to their activities as a publisher, e.g. ethics or diversity policies. When applied to a [[CreativeWork]] (e.g. [[NewsArticle]]) the principles are those of the party primarily responsible for the creation of the [[CreativeWork]].While such policies are most typically expressed in natural language, sometimes related information (e.g. indicating a [[funder]]) can be expressed using schema.org terminology.}

\item{parentOrganization}{(Organization type.) The larger organization that this organization is a [[subOrganization]] of, if any.}

\item{owns}{(Product or OwnershipInfo or Product or OwnershipInfo type.) Products owned by the organization or person.}

\item{numberOfEmployees}{(QuantitativeValue or QuantitativeValue type.) The number of employees in an organization e.g. business.}

\item{naics}{(Text or Text type.) The North American Industry Classification System (NAICS) code for a particular organization or business person.}

\item{members}{(Person or Organization or Person or Organization type.) A member of this organization.}

\item{memberOf}{(ProgramMembership or Organization or ProgramMembership or Organization type.) An Organization (or ProgramMembership) to which this Person or Organization belongs.}

\item{member}{(Person or Organization or Person or Organization type.) A member of an Organization or a ProgramMembership. Organizations can be members of organizations; ProgramMembership is typically for individuals.}

\item{makesOffer}{(Offer or Offer type.) A pointer to products or services offered by the organization or person.}

\item{logo}{(URL or ImageObject or URL or ImageObject or URL or ImageObject or URL or ImageObject or URL or ImageObject type.) An associated logo.}

\item{location}{(Text or PostalAddress or Place or Text or PostalAddress or Place or Text or PostalAddress or Place type.) The location of for example where the event is happening, an organization is located, or where an action takes place.}

\item{leiCode}{(Text type.) An organization identifier that uniquely identifies a legal entity as defined in ISO 17442.}

\item{legalName}{(Text type.) The official name of the organization, e.g. the registered company name.}

\item{isicV4}{(Text or Text or Text type.) The International Standard of Industrial Classification of All Economic Activities (ISIC), Revision 4 code for a particular organization, business person, or place.}

\item{hasPOS}{(Place or Place type.) Points-of-Sales operated by the organization or person.}

\item{hasOfferCatalog}{(OfferCatalog or OfferCatalog or OfferCatalog type.) Indicates an OfferCatalog listing for this Organization, Person, or Service.}

\item{globalLocationNumber}{(Text or Text or Text type.) The [Global Location Number](http://www.gs1.org/gln) (GLN, sometimes also referred to as International Location Number or ILN) of the respective organization, person, or place. The GLN is a 13-digit number used to identify parties and physical locations.}

\item{funder}{(Person or Organization or Person or Organization or Person or Organization or Person or Organization type.) A person or organization that supports (sponsors) something through some kind of financial contribution.}

\item{foundingLocation}{(Place type.) The place where the Organization was founded.}

\item{foundingDate}{(Date type.) The date that this organization was founded.}

\item{founders}{(Person type.) A person who founded this organization.}

\item{founder}{(Person type.) A person who founded this organization.}

\item{faxNumber}{(Text or Text or Text or Text type.) The fax number.}

\item{events}{(Event or Event type.) Upcoming or past events associated with this place or organization.}

\item{event}{(Event or Event or Event or Event or Event or Event or Event type.) Upcoming or past event associated with this place, organization, or action.}

\item{employees}{(Person type.) People working for this organization.}

\item{employee}{(Person type.) Someone working for this organization.}

\item{email}{(Text or Text or Text type.) Email address.}

\item{duns}{(Text or Text type.) The Dun & Bradstreet DUNS number for identifying an organization or business person.}

\item{dissolutionDate}{(Date type.) The date that this organization was dissolved.}

\item{department}{(Organization type.) A relationship between an organization and a department of that organization, also described as an organization (allowing different urls, logos, opening hours). For example: a store with a pharmacy, or a bakery with a cafe.}

\item{contactPoints}{(ContactPoint or ContactPoint type.) A contact point for a person or organization.}

\item{contactPoint}{(ContactPoint or ContactPoint type.) A contact point for a person or organization.}

\item{brand}{(Organization or Brand or Organization or Brand or Organization or Brand or Organization or Brand type.) The brand(s) associated with a product or service, or the brand(s) maintained by an organization or business person.}

\item{awards}{(Text or Text or Text or Text type.) Awards won by or for this item.}

\item{award}{(Text or Text or Text or Text or Text type.) An award won by or for this item.}

\item{areaServed}{(Text or Place or GeoShape or AdministrativeArea or Text or Place or GeoShape or AdministrativeArea or Text or Place or GeoShape or AdministrativeArea or Text or Place or GeoShape or AdministrativeArea or Text or Place or GeoShape or AdministrativeArea or Text or Place or GeoShape or AdministrativeArea type.) The geographic area where a service or offered item is provided.}

\item{aggregateRating}{(AggregateRating or AggregateRating or AggregateRating or AggregateRating or AggregateRating or AggregateRating or AggregateRating or AggregateRating type.) The overall rating, based on a collection of reviews or ratings, of the item.}

\item{address}{(Text or PostalAddress or Text or PostalAddress or Text or PostalAddress or Text or PostalAddress or Text or PostalAddress type.) Physical address of the item.}

\item{url}{(URL type.) URL of the item.}

\item{sameAs}{(URL type.) URL of a reference Web page that unambiguously indicates the item's identity. E.g. the URL of the item's Wikipedia page, Wikidata entry, or official website.}

\item{potentialAction}{(Action type.) Indicates a potential Action, which describes an idealized action in which this thing would play an 'object' role.}

\item{name}{(Text type.) The name of the item.}

\item{mainEntityOfPage}{(URL or CreativeWork type.) Indicates a page (or other CreativeWork) for which this thing is the main entity being described. See [background notes](/docs/datamodel.html#mainEntityBackground) for details.}

\item{image}{(URL or ImageObject type.) An image of the item. This can be a [[URL]] or a fully described [[ImageObject]].}

\item{identifier}{(URL or Text or PropertyValue type.) The identifier property represents any kind of identifier for any kind of [[Thing]], such as ISBNs, GTIN codes, UUIDs etc. Schema.org provides dedicated properties for representing many of these, either as textual strings or as URL (URI) links. See [background notes](/docs/datamodel.html#identifierBg) for more details.}

\item{disambiguatingDescription}{(Text type.) A sub property of description. A short description of the item used to disambiguate from other, similar items. Information from other properties (in particular, name) may be necessary for the description to be useful for disambiguation.}

\item{description}{(Text type.) A description of the item.}

\item{alternateName}{(Text type.) An alias for the item.}

\item{additionalType}{(URL type.) An additional type for the item, typically used for adding more specific types from external vocabularies in microdata syntax. This is a relationship between something and a class that the thing is in. In RDFa syntax, it is better to use the native RDFa syntax - the 'typeof' attribute - for multiple types. Schema.org tools may have only weaker understanding of extra types, in particular those defined externally.}
}
\value{
a list object corresponding to a schema:EducationalOrganization
}
\description{
An educational organization.
}
