% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExerciseAction.R
\name{ExerciseAction}
\alias{ExerciseAction}
\title{ExerciseAction}
\usage{
ExerciseAction(id = NULL, target = NULL, startTime = NULL,
  result = NULL, participant = NULL, object = NULL, location = NULL,
  instrument = NULL, error = NULL, endTime = NULL, agent = NULL,
  actionStatus = NULL, toLocation = NULL, sportsTeam = NULL,
  sportsEvent = NULL, sportsActivityLocation = NULL, opponent = NULL,
  fromLocation = NULL, exerciseCourse = NULL, distance = NULL,
  course = NULL, event = NULL, audience = NULL, url = NULL,
  sameAs = NULL, potentialAction = NULL, name = NULL,
  mainEntityOfPage = NULL, image = NULL, identifier = NULL,
  disambiguatingDescription = NULL, description = NULL,
  alternateName = NULL, additionalType = NULL)
}
\arguments{
\item{id}{identifier for the object (URI)}

\item{target}{(EntryPoint type.) Indicates a target EntryPoint for an Action.}

\item{startTime}{(DateTime or DateTime type.) The startTime of something. For a reserved event or service (e.g. FoodEstablishmentReservation), the time that it is expected to start. For actions that span a period of time, when the action was performed. e.g. John wrote a book from *January* to December.Note that Event uses startDate/endDate instead of startTime/endTime, even when describing dates with times. This situation may be clarified in future revisions.}

\item{result}{(Thing type.) The result produced in the action. e.g. John wrote *a book*.}

\item{participant}{(Person or Organization type.) Other co-agents that participated in the action indirectly. e.g. John wrote a book with *Steve*.}

\item{object}{(Thing type.) The object upon which the action is carried out, whose state is kept intact or changed. Also known as the semantic roles patient, affected or undergoer (which change their state) or theme (which doesn't). e.g. John read *a book*.}

\item{location}{(Text or PostalAddress or Place or Text or PostalAddress or Place or Text or PostalAddress or Place type.) The location of for example where the event is happening, an organization is located, or where an action takes place.}

\item{instrument}{(Thing type.) The object that helped the agent perform the action. e.g. John wrote a book with *a pen*.}

\item{error}{(Thing type.) For failed actions, more information on the cause of the failure.}

\item{endTime}{(DateTime or DateTime type.) The endTime of something. For a reserved event or service (e.g. FoodEstablishmentReservation), the time that it is expected to end. For actions that span a period of time, when the action was performed. e.g. John wrote a book from January to *December*.Note that Event uses startDate/endDate instead of startTime/endTime, even when describing dates with times. This situation may be clarified in future revisions.}

\item{agent}{(Person or Organization type.) The direct performer or driver of the action (animate or inanimate). e.g. *John* wrote a book.}

\item{actionStatus}{(ActionStatusType type.) Indicates the current disposition of the Action.}

\item{toLocation}{(Place or Place or Place or Place type.) A sub property of location. The final location of the object or the agent after the action.}

\item{sportsTeam}{(SportsTeam type.) A sub property of participant. The sports team that participated on this action.}

\item{sportsEvent}{(SportsEvent type.) A sub property of location. The sports event where this action occurred.}

\item{sportsActivityLocation}{(SportsActivityLocation type.) A sub property of location. The sports activity location where this action occurred.}

\item{opponent}{(Person type.) A sub property of participant. The opponent on this action.}

\item{fromLocation}{(Place or Place or Place type.) A sub property of location. The original location of the object or the agent before the action.}

\item{exerciseCourse}{(Place type.) A sub property of location. The course where this action was taken.}

\item{distance}{(Distance or Distance type.) The distance travelled, e.g. exercising or travelling.}

\item{course}{(Place type.) A sub property of location. The course where this action was taken.}

\item{event}{(Event or Event or Event or Event or Event or Event or Event type.) Upcoming or past event associated with this place, organization, or action.}

\item{audience}{(Audience or Audience or Audience or Audience or Audience or Audience type.) An intended audience, i.e. a group for whom something was created.}

\item{url}{(URL type.) URL of the item.}

\item{sameAs}{(URL type.) URL of a reference Web page that unambiguously indicates the item's identity. E.g. the URL of the item's Wikipedia page, Wikidata entry, or official website.}

\item{potentialAction}{(Action type.) Indicates a potential Action, which describes an idealized action in which this thing would play an 'object' role.}

\item{name}{(Text type.) The name of the item.}

\item{mainEntityOfPage}{(URL or CreativeWork type.) Indicates a page (or other CreativeWork) for which this thing is the main entity being described. See [background notes](/docs/datamodel.html#mainEntityBackground) for details.}

\item{image}{(URL or ImageObject type.) An image of the item. This can be a [[URL]] or a fully described [[ImageObject]].}

\item{identifier}{(URL or Text or PropertyValue type.) The identifier property represents any kind of identifier for any kind of [[Thing]], such as ISBNs, GTIN codes, UUIDs etc. Schema.org provides dedicated properties for representing many of these, either as textual strings or as URL (URI) links. See [background notes](/docs/datamodel.html#identifierBg) for more details.}

\item{disambiguatingDescription}{(Text type.) A sub property of description. A short description of the item used to disambiguate from other, similar items. Information from other properties (in particular, name) may be necessary for the description to be useful for disambiguation.}

\item{description}{(Text type.) A description of the item.}

\item{alternateName}{(Text type.) An alias for the item.}

\item{additionalType}{(URL type.) An additional type for the item, typically used for adding more specific types from external vocabularies in microdata syntax. This is a relationship between something and a class that the thing is in. In RDFa syntax, it is better to use the native RDFa syntax - the 'typeof' attribute - for multiple types. Schema.org tools may have only weaker understanding of extra types, in particular those defined externally.}
}
\value{
a list object corresponding to a schema:ExerciseAction
}
\description{
The act of participating in exertive activity for the purposes of improving health and fitness.
}
