% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\name{getSchema}
\alias{getSchema}
\title{get schema from data.frame}
\usage{
getSchema(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
list of colnames and their correspondent types
}
\description{
it's only for help purpose. You still need to convert the result
to "list" function call and you would probably do it by hand :(
In some cases you can use that as a snapshot and don't write it down... but
it wouldn't be a good practice (you should always know the schema beforehand)
}
\examples{
getSchema(iris)

}
