% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\name{isThisDFOk}
\alias{isThisDFOk}
\title{Create check function}
\usage{
isThisDFOk(schema)
}
\arguments{
\item{schema}{schema you want ot compare to}

\item{df}{data.frame you want to check}
}
\value{
function
}
\description{
this function create checker functions (it's mostly for readability)
}
\examples{
df = data.frame(foo = "x", bar = "y")
isFooBar = isThisDFOk(schema_example)
isFooBar(df)
}
