% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{makeItPassIf}
\alias{makeItPassIf}
\title{Enable pipe validations}
\usage{
makeItPassIf(fun)
}
\arguments{
\item{fun}{function created from isThisDFOk or other}
}
\value{
function that takes df as it's argument and checks it's validity from
function you've provided
}
\description{
Enable pipe validations
}
\examples{
thisIsIris = isThisDFOk(schema_iris)
df = makeItPassIf(thisIsIris)(iris)
head(df)

makeItPassIf(thisIsIris)(mtcars)
}
