% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Order.R
\name{Order}
\alias{Order}
\title{Order}
\usage{
Order(id = NULL, seller = NULL, paymentUrl = NULL,
  paymentMethodId = NULL, paymentMethod = NULL, paymentDueDate = NULL,
  paymentDue = NULL, partOfInvoice = NULL, orderedItem = NULL,
  orderStatus = NULL, orderNumber = NULL, orderDelivery = NULL,
  orderDate = NULL, merchant = NULL, isGift = NULL,
  discountCurrency = NULL, discountCode = NULL, discount = NULL,
  customer = NULL, confirmationNumber = NULL, broker = NULL,
  billingAddress = NULL, acceptedOffer = NULL, url = NULL,
  sameAs = NULL, potentialAction = NULL, name = NULL,
  mainEntityOfPage = NULL, image = NULL, identifier = NULL,
  disambiguatingDescription = NULL, description = NULL,
  alternateName = NULL, additionalType = NULL)
}
\arguments{
\item{id}{identifier for the object (URI)}

\item{seller}{(Person or Organization or Person or Organization or Person or Organization or Person or Organization or Person or Organization type.) An entity which offers (sells / leases / lends / loans) the services / goods.  A seller may also be a provider.}

\item{paymentUrl}{(URL type.) The URL for sending a payment.}

\item{paymentMethodId}{(Text or Text type.) An identifier for the method of payment used (e.g. the last 4 digits of the credit card).}

\item{paymentMethod}{(PaymentMethod or PaymentMethod type.) The name of the credit card or other method of payment for the order.}

\item{paymentDueDate}{(DateTime or DateTime type.) The date that payment is due.}

\item{paymentDue}{(DateTime or DateTime type.) The date that payment is due.}

\item{partOfInvoice}{(Invoice type.) The order is being paid as part of the referenced Invoice.}

\item{orderedItem}{(Product or OrderItem or Product or OrderItem type.) The item ordered.}

\item{orderStatus}{(OrderStatus type.) The current status of the order.}

\item{orderNumber}{(Text type.) The identifier of the transaction.}

\item{orderDelivery}{(ParcelDelivery or ParcelDelivery type.) The delivery of the parcel related to this order or order item.}

\item{orderDate}{(DateTime type.) Date order was placed.}

\item{merchant}{(Person or Organization type.) 'merchant' is an out-dated term for 'seller'.}

\item{isGift}{(Boolean type.) Was the offer accepted as a gift for someone other than the buyer.}

\item{discountCurrency}{(Text type.) The currency of the discount.Use standard formats: [ISO 4217 currency format](http://en.wikipedia.org/wiki/ISO_4217) e.g. "USD"; [Ticker symbol](https://en.wikipedia.org/wiki/List_of_cryptocurrencies) for cryptocurrencies e.g. "BTC"; well known names for [Local Exchange Tradings Systems](https://en.wikipedia.org/wiki/Local_exchange_trading_system) (LETS) and other currency types e.g. "Ithaca HOUR".}

\item{discountCode}{(Text type.) Code used to redeem a discount.}

\item{discount}{(Text or Number type.) Any discount applied (to an Order).}

\item{customer}{(Person or Organization or Person or Organization type.) Party placing the order or paying the invoice.}

\item{confirmationNumber}{(Text or Text type.) A number that confirms the given order or payment has been received.}

\item{broker}{(Person or Organization or Person or Organization or Person or Organization or Person or Organization type.) An entity that arranges for an exchange between a buyer and a seller.  In most cases a broker never acquires or releases ownership of a product or service involved in an exchange.  If it is not clear whether an entity is a broker, seller, or buyer, the latter two terms are preferred.}

\item{billingAddress}{(PostalAddress type.) The billing address for the order.}

\item{acceptedOffer}{(Offer type.) The offer(s) -- e.g., product, quantity and price combinations -- included in the order.}

\item{url}{(URL type.) URL of the item.}

\item{sameAs}{(URL type.) URL of a reference Web page that unambiguously indicates the item's identity. E.g. the URL of the item's Wikipedia page, Wikidata entry, or official website.}

\item{potentialAction}{(Action type.) Indicates a potential Action, which describes an idealized action in which this thing would play an 'object' role.}

\item{name}{(Text type.) The name of the item.}

\item{mainEntityOfPage}{(URL or CreativeWork type.) Indicates a page (or other CreativeWork) for which this thing is the main entity being described. See [background notes](/docs/datamodel.html#mainEntityBackground) for details.}

\item{image}{(URL or ImageObject type.) An image of the item. This can be a [[URL]] or a fully described [[ImageObject]].}

\item{identifier}{(URL or Text or PropertyValue type.) The identifier property represents any kind of identifier for any kind of [[Thing]], such as ISBNs, GTIN codes, UUIDs etc. Schema.org provides dedicated properties for representing many of these, either as textual strings or as URL (URI) links. See [background notes](/docs/datamodel.html#identifierBg) for more details.}

\item{disambiguatingDescription}{(Text type.) A sub property of description. A short description of the item used to disambiguate from other, similar items. Information from other properties (in particular, name) may be necessary for the description to be useful for disambiguation.}

\item{description}{(Text type.) A description of the item.}

\item{alternateName}{(Text type.) An alias for the item.}

\item{additionalType}{(URL type.) An additional type for the item, typically used for adding more specific types from external vocabularies in microdata syntax. This is a relationship between something and a class that the thing is in. In RDFa syntax, it is better to use the native RDFa syntax - the 'typeof' attribute - for multiple types. Schema.org tools may have only weaker understanding of extra types, in particular those defined externally.}
}
\value{
a list object corresponding to a schema:Order
}
\description{
An order is a confirmation of a transaction (a receipt), which can contain multiple line items, each represented by an Offer that has been accepted by the customer.
}
