% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParcelDelivery.R
\name{ParcelDelivery}
\alias{ParcelDelivery}
\title{ParcelDelivery}
\usage{
ParcelDelivery(id = NULL, trackingUrl = NULL, trackingNumber = NULL,
  provider = NULL, partOfOrder = NULL, originAddress = NULL,
  itemShipped = NULL, hasDeliveryMethod = NULL,
  expectedArrivalUntil = NULL, expectedArrivalFrom = NULL,
  deliveryStatus = NULL, deliveryAddress = NULL, carrier = NULL,
  url = NULL, sameAs = NULL, potentialAction = NULL, name = NULL,
  mainEntityOfPage = NULL, image = NULL, identifier = NULL,
  disambiguatingDescription = NULL, description = NULL,
  alternateName = NULL, additionalType = NULL)
}
\arguments{
\item{id}{identifier for the object (URI)}

\item{trackingUrl}{(URL type.) Tracking url for the parcel delivery.}

\item{trackingNumber}{(Text type.) Shipper tracking number.}

\item{provider}{(Person or Organization or Person or Organization or Person or Organization or Person or Organization or Person or Organization or Person or Organization type.) The service provider, service operator, or service performer; the goods producer. Another party (a seller) may offer those services or goods on behalf of the provider. A provider may also serve as the seller.}

\item{partOfOrder}{(Order type.) The overall order the items in this delivery were included in.}

\item{originAddress}{(PostalAddress type.) Shipper's address.}

\item{itemShipped}{(Product type.) Item(s) being shipped.}

\item{hasDeliveryMethod}{(DeliveryMethod or DeliveryMethod type.) Method used for delivery or shipping.}

\item{expectedArrivalUntil}{(DateTime type.) The latest date the package may arrive.}

\item{expectedArrivalFrom}{(DateTime type.) The earliest date the package may arrive.}

\item{deliveryStatus}{(DeliveryEvent type.) New entry added as the package passes through each leg of its journey (from shipment to final delivery).}

\item{deliveryAddress}{(PostalAddress type.) Destination address.}

\item{carrier}{(Organization or Organization type.) 'carrier' is an out-dated term indicating the 'provider' for parcel delivery and flights.}

\item{url}{(URL type.) URL of the item.}

\item{sameAs}{(URL type.) URL of a reference Web page that unambiguously indicates the item's identity. E.g. the URL of the item's Wikipedia page, Wikidata entry, or official website.}

\item{potentialAction}{(Action type.) Indicates a potential Action, which describes an idealized action in which this thing would play an 'object' role.}

\item{name}{(Text type.) The name of the item.}

\item{mainEntityOfPage}{(URL or CreativeWork type.) Indicates a page (or other CreativeWork) for which this thing is the main entity being described. See [background notes](/docs/datamodel.html#mainEntityBackground) for details.}

\item{image}{(URL or ImageObject type.) An image of the item. This can be a [[URL]] or a fully described [[ImageObject]].}

\item{identifier}{(URL or Text or PropertyValue type.) The identifier property represents any kind of identifier for any kind of [[Thing]], such as ISBNs, GTIN codes, UUIDs etc. Schema.org provides dedicated properties for representing many of these, either as textual strings or as URL (URI) links. See [background notes](/docs/datamodel.html#identifierBg) for more details.}

\item{disambiguatingDescription}{(Text type.) A sub property of description. A short description of the item used to disambiguate from other, similar items. Information from other properties (in particular, name) may be necessary for the description to be useful for disambiguation.}

\item{description}{(Text type.) A description of the item.}

\item{alternateName}{(Text type.) An alias for the item.}

\item{additionalType}{(URL type.) An additional type for the item, typically used for adding more specific types from external vocabularies in microdata syntax. This is a relationship between something and a class that the thing is in. In RDFa syntax, it is better to use the native RDFa syntax - the 'typeof' attribute - for multiple types. Schema.org tools may have only weaker understanding of extra types, in particular those defined externally.}
}
\value{
a list object corresponding to a schema:ParcelDelivery
}
\description{
The delivery of a parcel either via the postal service or a commercial service.
}
