% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{sch_schema}
\alias{sch_schema}
\alias{sch_others}
\alias{sch_nest}
\alias{sch_numeric}
\alias{sch_integer}
\alias{sch_logical}
\alias{sch_character}
\alias{sch_factor}
\alias{sch_date}
\alias{sch_datetime}
\alias{sch_inherits}
\alias{sch_list_of}
\alias{sch_custom}
\title{Define a structured data type}
\usage{
sch_schema(..., .desc = NULL)

sch_others()

sch_nest(..., .keys = character(0), .desc = NULL)

sch_numeric(
  desc = NULL,
  bounds = c(-Inf, Inf),
  closed = c(TRUE, TRUE),
  missing = TRUE,
  required = TRUE,
  distinct = FALSE
)

sch_integer(
  desc = NULL,
  bounds = c(-Inf, Inf),
  closed = c(TRUE, TRUE),
  missing = TRUE,
  required = TRUE,
  distinct = FALSE
)

sch_logical(desc = NULL, missing = TRUE, required = TRUE, distinct = FALSE)

sch_character(desc = NULL, missing = TRUE, required = TRUE, distinct = FALSE)

sch_factor(
  desc = NULL,
  levels,
  strict = TRUE,
  missing = TRUE,
  required = TRUE,
  distinct = FALSE
)

sch_date(
  desc = NULL,
  bounds = c(as.Date(-Inf), as.Date(Inf)),
  closed = c(FALSE, FALSE),
  missing = TRUE,
  required = TRUE,
  distinct = FALSE
)

sch_datetime(
  desc = NULL,
  bounds = c(as.POSIXct(-Inf), as.POSIXct(Inf)),
  closed = c(FALSE, FALSE),
  missing = TRUE,
  required = TRUE,
  distinct = FALSE
)

sch_inherits(
  desc = NULL,
  class,
  missing = TRUE,
  required = TRUE,
  distinct = FALSE
)

sch_list_of(
  desc = NULL,
  class,
  missing = TRUE,
  required = TRUE,
  distinct = FALSE
)

sch_custom(
  name,
  desc = NULL,
  check,
  msg,
  coerce,
  ...,
  missing = TRUE,
  required = TRUE,
  distinct = FALSE
)
}
\arguments{
\item{...}{Column specifications, in the form of \code{col_name = col_type} pairs,
where \code{col_type} is a call to a column type constructor listed here, such
as \code{sch_numeric()}. Every type must be a kind of vector, i.e.,
\code{\link[vctrs:vector-checks]{vctrs::obj_is_vector()}} must return \code{TRUE}.

All columns must be named, except for \code{sch_others()}, as described below,
and unnamed \code{sch_nest()}, which describes a set of columns which are
logically nested within the outer columns but are stored flat in the actual
data frame. A named \code{sch_nest()} describes columns stored as a nested
data frame.

The special function \code{sch_others()} indicates the preferred location of
other columns not explicitly mentioned in the schema. If no \code{sch_others()}
appears, then other columns are not allowed.
Trailing commas are permitted.}

\item{.keys}{A character vector selecting one or more column names from \code{...}
that serve as the key columns for the nested group.}

\item{desc, .desc}{A description of the column for consumers of the schema.
The type contraints will be described separately and do not need to be
included in the description.  For example for "age", the description might
be "Age of the patient in years", not "Non-negative integer representing
the age of the patient in years".}

\item{bounds}{Length-two vector \code{c(min, max)} specifying the allowed range of values.}

\item{closed}{Length-two logical vector specifying whether the bounds are
closed (inclusive) or open (exclusive).}

\item{missing}{If \code{TRUE}, the column may be contain missing values. Otherwise,
any missing values result in an error.}

\item{required}{If \code{TRUE}, the column must be present. If \code{FALSE}, the column
is optional.}

\item{distinct}{If \code{TRUE}, the column must contain no duplicate values (after
accounting for nesting structure).}

\item{levels}{A character vector of factor levels.}

\item{strict}{If \code{TRUE}, only factors with the specified levels are accepted.
If \code{FALSE}, character vectors with the specified levels are also accepted.}

\item{class}{A character vector of class names.}

\item{name}{A name for the custom type.}

\item{check}{A two-argument function that checks whether an object satisfies
the type. The first argument is the object to check, and the second is the
full type specification.}

\item{msg}{A one-argument function that generates a descriptive message
about the type when passed the type object itself. Should not end with a period.}

\item{coerce}{A two-argument function that attempts to coerce an object to the
type. The first argument is the object to coerce, and the second is the full
type specification.}
}
\value{
An object of class \code{sch_schema},
}
\description{
Defines the structure of a single 'observation' for a structured data frame.
Each column has type restrictions and may be required or optional.
Schemas support nesting relationships.
}
\section{Functions}{
\itemize{
\item \code{sch_others()}: A placeholder for other non-required columns in a schema.

\item \code{sch_nest()}: A set of columns that are logically nested within the
outer schema. If given a name in the outer \code{sch_schema()}, the columns are
stored as a nested data frame. If unnamed, the columns are stored flat
(adjacent to the outer columns). The unique combinations of rows defined
by \code{.keys} must repeat identically across groups defined by the outer columns.

\item \code{sch_numeric()}: A numeric vector that is optionally constrained to be
within a certain range.

\item \code{sch_integer()}: An integer vector that is optionally constrained to be
within a certain range.

\item \code{sch_logical()}: A logical vector.

\item \code{sch_character()}: A character vector.

\item \code{sch_factor()}: A factor with specified levels.

\item \code{sch_date()}: A Date vector that is optionally constrained to be
within a certain range.

\item \code{sch_datetime()}: A POSIXct vector that is optionally constrained to be
within a certain range.

\item \code{sch_inherits()}: A list-column whose elements satisfy \verb{inherits(_, class)}.

\item \code{sch_list_of()}: A vector satisfying \verb{inherits(_, class)}.

\item \code{sch_custom()}: A custom type defined by user-provided check, type message,
and coercion functions. Additional named values to be stored along with the
type specification may be passed via \code{...} and will be available to the
check, message, and coercion function as elements of the \code{type} argument.

}}
\examples{
sch_schema(
    .desc = "Student data",
    age = sch_integer("Age in years", bounds = c(0, 130)),
    birthday = sch_date("Date of birth", required = FALSE),
    height = sch_numeric(
        "Height in inches",
        bounds = c(0, 108),
        closed = c(FALSE, TRUE)
    ),
    teacher = sch_factor(levels = c("Jones", "Smith", "Hernandez")),
    enrolled = sch_logical(missing = FALSE),
    sch_others()
)

sch_schema(
    .desc = "MCMC draws",
    draw = sch_integer(
        "Draw number",
        bounds = c(1, Inf),
        closed = c(TRUE, FALSE),
        distinct = TRUE
    ),
    sch_nest(
        param = sch_factor("Parameter name", levels = c("mu", "sigma", "log_lik")),
        value = sch_numeric("Parameter value"),
        .keys = "param"
    )
)

sch_custom(
   name = "even",
   check = function(x, type) is.integer(x) && all(x \%\% 2 == 0),
   msg = function(type) "vector of even integers",
   coerce = function(x, type) (as.integer(x) \%/\% 2) * 2
)
}
