% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schmear-package.R
\docType{package}
\name{schmear-package}
\alias{schmear}
\alias{schmear-package}
\title{schmear: Build Structured Data Frame Subtypes}
\description{
Provides developer-focused helper functions and S3 classes to ease the creation of structured subtypes of data frames which require certain columns and types to be present, and which preserve special attributes through common data frame manipulations.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://corymccartan.com/schmear/}
}

}
\author{
\strong{Maintainer}: Cory McCartan \email{mccartan@psu.edu} (\href{https://orcid.org/0000-0002-6251-669X}{ORCID})

}
\keyword{internal}
