% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Service.R
\name{Service}
\alias{Service}
\title{Service}
\usage{
Service(id = NULL, serviceType = NULL, serviceOutput = NULL,
  serviceAudience = NULL, serviceArea = NULL, review = NULL,
  providerMobility = NULL, provider = NULL, produces = NULL,
  offers = NULL, logo = NULL, isSimilarTo = NULL, isRelatedTo = NULL,
  hoursAvailable = NULL, hasOfferCatalog = NULL, category = NULL,
  broker = NULL, brand = NULL, award = NULL, availableChannel = NULL,
  audience = NULL, areaServed = NULL, aggregateRating = NULL,
  url = NULL, sameAs = NULL, potentialAction = NULL, name = NULL,
  mainEntityOfPage = NULL, image = NULL, identifier = NULL,
  disambiguatingDescription = NULL, description = NULL,
  alternateName = NULL, additionalType = NULL)
}
\arguments{
\item{id}{identifier for the object (URI)}

\item{serviceType}{(Text type.) The type of service being offered, e.g. veterans' benefits, emergency relief, etc.}

\item{serviceOutput}{(Thing type.) The tangible thing generated by the service, e.g. a passport, permit, etc.}

\item{serviceAudience}{(Audience type.) The audience eligible for this service.}

\item{serviceArea}{(Place or GeoShape or AdministrativeArea or Place or GeoShape or AdministrativeArea or Place or GeoShape or AdministrativeArea type.) The geographic area where the service is provided.}

\item{review}{(Review or Review or Review or Review or Review or Review or Review or Review type.) A review of the item.}

\item{providerMobility}{(Text type.) Indicates the mobility of a provided service (e.g. 'static', 'dynamic').}

\item{provider}{(Person or Organization or Person or Organization or Person or Organization or Person or Organization or Person or Organization or Person or Organization type.) The service provider, service operator, or service performer; the goods producer. Another party (a seller) may offer those services or goods on behalf of the provider. A provider may also serve as the seller.}

\item{produces}{(Thing type.) The tangible thing generated by the service, e.g. a passport, permit, etc.}

\item{offers}{(Offer or Offer or Offer or Offer or Offer or Offer or Offer type.) An offer to provide this item&#x2014;for example, an offer to sell a product, rent the DVD of a movie, perform a service, or give away tickets to an event.}

\item{logo}{(URL or ImageObject or URL or ImageObject or URL or ImageObject or URL or ImageObject or URL or ImageObject type.) An associated logo.}

\item{isSimilarTo}{(Service or Product or Service or Product type.) A pointer to another, functionally similar product (or multiple products).}

\item{isRelatedTo}{(Service or Product or Service or Product type.) A pointer to another, somehow related product (or multiple products).}

\item{hoursAvailable}{(OpeningHoursSpecification or OpeningHoursSpecification or OpeningHoursSpecification type.) The hours during which this service or contact is available.}

\item{hasOfferCatalog}{(OfferCatalog or OfferCatalog or OfferCatalog type.) Indicates an OfferCatalog listing for this Organization, Person, or Service.}

\item{category}{(Thing or Text or Thing or Text or Thing or Text or Thing or Text type.) A category for the item. Greater signs or slashes can be used to informally indicate a category hierarchy.}

\item{broker}{(Person or Organization or Person or Organization or Person or Organization or Person or Organization type.) An entity that arranges for an exchange between a buyer and a seller.  In most cases a broker never acquires or releases ownership of a product or service involved in an exchange.  If it is not clear whether an entity is a broker, seller, or buyer, the latter two terms are preferred.}

\item{brand}{(Organization or Brand or Organization or Brand or Organization or Brand or Organization or Brand type.) The brand(s) associated with a product or service, or the brand(s) maintained by an organization or business person.}

\item{award}{(Text or Text or Text or Text or Text type.) An award won by or for this item.}

\item{availableChannel}{(ServiceChannel type.) A means of accessing the service (e.g. a phone bank, a web site, a location, etc.).}

\item{audience}{(Audience or Audience or Audience or Audience or Audience or Audience type.) An intended audience, i.e. a group for whom something was created.}

\item{areaServed}{(Text or Place or GeoShape or AdministrativeArea or Text or Place or GeoShape or AdministrativeArea or Text or Place or GeoShape or AdministrativeArea or Text or Place or GeoShape or AdministrativeArea or Text or Place or GeoShape or AdministrativeArea or Text or Place or GeoShape or AdministrativeArea type.) The geographic area where a service or offered item is provided.}

\item{aggregateRating}{(AggregateRating or AggregateRating or AggregateRating or AggregateRating or AggregateRating or AggregateRating or AggregateRating or AggregateRating type.) The overall rating, based on a collection of reviews or ratings, of the item.}

\item{url}{(URL type.) URL of the item.}

\item{sameAs}{(URL type.) URL of a reference Web page that unambiguously indicates the item's identity. E.g. the URL of the item's Wikipedia page, Wikidata entry, or official website.}

\item{potentialAction}{(Action type.) Indicates a potential Action, which describes an idealized action in which this thing would play an 'object' role.}

\item{name}{(Text type.) The name of the item.}

\item{mainEntityOfPage}{(URL or CreativeWork type.) Indicates a page (or other CreativeWork) for which this thing is the main entity being described. See [background notes](/docs/datamodel.html#mainEntityBackground) for details.}

\item{image}{(URL or ImageObject type.) An image of the item. This can be a [[URL]] or a fully described [[ImageObject]].}

\item{identifier}{(URL or Text or PropertyValue type.) The identifier property represents any kind of identifier for any kind of [[Thing]], such as ISBNs, GTIN codes, UUIDs etc. Schema.org provides dedicated properties for representing many of these, either as textual strings or as URL (URI) links. See [background notes](/docs/datamodel.html#identifierBg) for more details.}

\item{disambiguatingDescription}{(Text type.) A sub property of description. A short description of the item used to disambiguate from other, similar items. Information from other properties (in particular, name) may be necessary for the description to be useful for disambiguation.}

\item{description}{(Text type.) A description of the item.}

\item{alternateName}{(Text type.) An alias for the item.}

\item{additionalType}{(URL type.) An additional type for the item, typically used for adding more specific types from external vocabularies in microdata syntax. This is a relationship between something and a class that the thing is in. In RDFa syntax, it is better to use the native RDFa syntax - the 'typeof' attribute - for multiple types. Schema.org tools may have only weaker understanding of extra types, in particular those defined externally.}
}
\value{
a list object corresponding to a schema:Service
}
\description{
A service provided by an organization, e.g. delivery service, print services, etc.
}
