% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TypeAndQuantityNode.R
\name{TypeAndQuantityNode}
\alias{TypeAndQuantityNode}
\title{TypeAndQuantityNode}
\usage{
TypeAndQuantityNode(id = NULL, url = NULL, sameAs = NULL,
  potentialAction = NULL, name = NULL, mainEntityOfPage = NULL,
  image = NULL, identifier = NULL, disambiguatingDescription = NULL,
  description = NULL, alternateName = NULL, additionalType = NULL,
  unitText = NULL, unitCode = NULL, typeOfGood = NULL,
  businessFunction = NULL, amountOfThisGood = NULL)
}
\arguments{
\item{id}{identifier for the object (URI)}

\item{url}{(URL type.) URL of the item.}

\item{sameAs}{(URL type.) URL of a reference Web page that unambiguously indicates the item's identity. E.g. the URL of the item's Wikipedia page, Wikidata entry, or official website.}

\item{potentialAction}{(Action type.) Indicates a potential Action, which describes an idealized action in which this thing would play an 'object' role.}

\item{name}{(Text type.) The name of the item.}

\item{mainEntityOfPage}{(URL or CreativeWork type.) Indicates a page (or other CreativeWork) for which this thing is the main entity being described. See [background notes](/docs/datamodel.html#mainEntityBackground) for details.}

\item{image}{(URL or ImageObject type.) An image of the item. This can be a [[URL]] or a fully described [[ImageObject]].}

\item{identifier}{(URL or Text or PropertyValue type.) The identifier property represents any kind of identifier for any kind of [[Thing]], such as ISBNs, GTIN codes, UUIDs etc. Schema.org provides dedicated properties for representing many of these, either as textual strings or as URL (URI) links. See [background notes](/docs/datamodel.html#identifierBg) for more details.}

\item{disambiguatingDescription}{(Text type.) A sub property of description. A short description of the item used to disambiguate from other, similar items. Information from other properties (in particular, name) may be necessary for the description to be useful for disambiguation.}

\item{description}{(Text type.) A description of the item.}

\item{alternateName}{(Text type.) An alias for the item.}

\item{additionalType}{(URL type.) An additional type for the item, typically used for adding more specific types from external vocabularies in microdata syntax. This is a relationship between something and a class that the thing is in. In RDFa syntax, it is better to use the native RDFa syntax - the 'typeof' attribute - for multiple types. Schema.org tools may have only weaker understanding of extra types, in particular those defined externally.}

\item{unitText}{(Text or Text or Text or Text type.) A string or text indicating the unit of measurement. Useful if you cannot provide a standard unit code for<a href='unitCode'>unitCode</a>.}

\item{unitCode}{(URL or Text or URL or Text or URL or Text or URL or Text type.) The unit of measurement given using the UN/CEFACT Common Code (3 characters) or a URL. Other codes than the UN/CEFACT Common Code may be used with a prefix followed by a colon.}

\item{typeOfGood}{(Service or Product or Service or Product type.) The product that this structured value is referring to.}

\item{businessFunction}{(BusinessFunction or BusinessFunction or BusinessFunction type.) The business function (e.g. sell, lease, repair, dispose) of the offer or component of a bundle (TypeAndQuantityNode). The default is http://purl.org/goodrelations/v1#Sell.}

\item{amountOfThisGood}{(Number type.) The quantity of the goods included in the offer.}
}
\value{
a list object corresponding to a schema:TypeAndQuantityNode
}
\description{
A structured value indicating the quantity, unit of measurement, and business function of goods included in a bundle offer.
}
